Fixups

+++ publish.sh
@@ -5,6 +5,10 @@ if [[ -z "$commit_msg" ]]; then
   echo "Empty commit message, aborting" >&2
   exit 1
 fi
+if [[ ! -z "$2" ]]; then
+  echo "Too many arguments" >&2
+  exit 1
+fi
 
 rm -rf ./public/*
 hugo --minify || exit 1
